gio/gcontenttype-fdo: Do not overflow if header is longer than MAXINT
In case the header size is longer than MAXINT we may read and write to
invalid locations

Spotted by treeplus.
Thanks to the Sovereign Tech Resilience programme from the Sovereign
Tech Agency.

ID: #YWH-PGM9867-169
Closes: #3871

+++ gio/gcontenttype-fdo.c
@@ -817,7 +817,7 @@ tree_match_free (TreeMatch *match)
 static TreeMatch *
 parse_header (gchar *line)
 {
-  gint len;
+  size_t len;
   gchar *s;
   TreeMatch *match;
